+++ lib/content-types.js
@@ -5,7 +5,7 @@ const _ = require('lodash');
 const loader = require('./content-types/load');
 const merge = require('./content-types/merge');
 const only = require('./content-types/only');
-const util = require('./util');
+const utils = require('./utils');
 
 /**
  * Get single Content Type
@@ -16,7 +16,7 @@ const util = require('./util');
  * @returns {object} content type object
  */
 const findCT = (id, types) => {
-  const ct = util.singleItem('id', id, types);
+  const ct = utils.singleItem('id', id, types);
 
   if (ct === false) {
     throw new Error(`Content type ${id} not found `);
