+++ lib/content-types/load.js
@@ -5,7 +5,7 @@ const yaml = require('js-yaml');
 const slugify = require('lodash/kebabCase');
 const globalConfig = require('config');
 const path = require('path');
-const util = require('../util');
+const utils = require('../utils');
 
 
 /**
@@ -111,7 +111,7 @@ const names = types => {
  * @returns {array} array of content type's attribute objects
  */
 const attributes = (id, types) => {
-  const filtered = util.singleItem('id', id, types);
+  const filtered = utils.singleItem('id', id, types);
 
   return filtered ? filtered[0].attributes : false;
 };
