+++ lib/content-types/only.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const merge = require('deepmerge');
-const util = require('../util');
+const utils = require('../utils');
 
 const only = (type, config) => {
   return new Promise((resolve) => {
@@ -30,8 +30,8 @@ const only = (type, config) => {
           inputs.forEach(input => {
             instance[input].validation = attribute.inputs[0][input].validation;
             instance[input].type = attribute.inputs[0][input].type;
-            instance[input].id = `${util.splitPop(attribute.inputs[0][input].id, '--')}--${index}`;
-            instance[input].name = `${util.splitPop(attribute.inputs[0][input].name, '--')}--${index}`;
+            instance[input].id = `${utils.splitPop(attribute.inputs[0][input].id, '--')}--${index}`;
+            instance[input].name = `${utils.splitPop(attribute.inputs[0][input].name, '--')}--${index}`;
           });
 
           return instance;
