+++ lib/form/scripts.js
@@ -5,7 +5,7 @@ const stream = require('stream');
 const fs = require('fs');
 const path = require('path');
 const _ = require('lodash');
-const utils = require('../util');
+const utils = require('../utils');
 const globalConfig = require('config');
 const formJS = fs.readdirSync(path.join(__dirname, 'js')).map(script => {
   return fs.readFileSync(path.join(__dirname, 'js', script));
