+++ lib/form/validate.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const util = require('../util');
+const utils = require('../utils');
 
 /*
  * @typedef FormInputValues
@@ -230,7 +230,7 @@ const validate = (raw, type) => {
   const values = buildValues(working);
   working = working.map(value => {
     const result = value;
-    const plugin = util.singleItem('id', value.plugin, type.attributes);
+    const plugin = utils.singleItem('id', value.plugin, type.attributes);
     const input = {};
     const settings = {};
